Refactor: Fileio => GetMethod
