+++ include/core/process.hpp
@@ -3,7 +3,7 @@
 
 #include "Server.hpp"
 #include "Client.hpp"
-#include "Fileio.hpp"
+#include "GetMethod.hpp"
 #include "NotFoundError.hpp"
 
 void Process(FdInterface *target, struct kevent event);
