+++ include/util/utils.hpp
@@ -5,6 +5,8 @@
 #include <vector>
 #include <iostream>
 #include <sstream>
+#include <fcntl.h>
+#include <unistd.h>
 
 void ExitWithPerror(const std::string &msg);
 std::string CheckArg(int argc, char **argv);
@@ -15,6 +17,7 @@ std::string& rtrim(std::string& s);
 std::string& trim(std::string& s);
 std::string ft_itos(int i);
 int ft_stoi(const std::string &s);
+int IsEOF(int fd);
 
 std::string StatusCode(const int &code);
 std::string DefaultErrorPage(const int &code);
