+++ src/main.cpp
@@ -2,7 +2,7 @@
 #include "ServerManager.hpp"
 #include "KQueue.hpp"
 #include "Server.hpp"
-#include "Fileio.hpp"
+#include "GetMethod.hpp"
 #include "process.hpp"
 
 int main(int argc, char** argv) {
